IS160_Project-main/IS160_Project-main/webapp/requirements.txt
Flask==3.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
Werkzeug==3.0.1

